refactor: move definitions from qn to solving
